+++ src/image-viewer/_example/base.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import {ImageViewer, Image, Space} from 'tdesign-react';
-import {BrowseIcon} from 'tdesign-icons-react'
+import { ImageViewer, Image, Space } from 'tdesign-react';
+import { BrowseIcon } from 'tdesign-icons-react';
 
 const img = 'https://tdesign.gtimg.com/demo/demo-image-1.png';
 
 export default function BasicImageViewer() {
-  const trigger = ({open}) => {
+  const trigger = ({ open }) => {
     const mask = (
       <div
         style={{
@@ -14,11 +14,13 @@ export default function BasicImageViewer() {
           height: '100%',
           display: 'flex',
           alignItems: 'center',
-          justifyContent: 'center'
+          justifyContent: 'center',
         }}
         onClick={open}
       >
-        <span><BrowseIcon size="16px" name={'browse'}/> 预览</span>
+        <span>
+          <BrowseIcon size="16px" name={'browse'} /> 预览
+        </span>
       </div>
     );
 
@@ -34,15 +36,15 @@ export default function BasicImageViewer() {
           height: 160,
           border: '4px solid var(--td-bg-color-secondarycontainer)',
           borderRadius: 'var(--td-radius-medium)',
-          backgroundColor: '#fff'
+          backgroundColor: '#fff',
         }}
       />
-    )
-  }
+    );
+  };
 
   return (
     <Space breakLine size={16}>
-      <ImageViewer trigger={trigger} images={[img]}/>
+      <ImageViewer trigger={trigger} images={[img]} />
 
       {/* TODO: fix visible=true can not show image previewer */}
       {/* <ImageViewer images={[img]} visible={true} /> */}