+++ src/table/BaseTable.tsx
@@ -1,10 +1,19 @@
-import React, { useRef, useMemo, useImperativeHandle, forwardRef, useEffect, useState, WheelEvent } from 'react';
+import React, {
+  useRef,
+  useMemo,
+  useImperativeHandle,
+  forwardRef,
+  useEffect,
+  useState,
+  WheelEvent,
+  RefAttributes,
+} from 'react';
 import pick from 'lodash/pick';
 import classNames from 'classnames';
 import TBody, { extendTableProps, TableBodyProps } from './TBody';
-import { Affix } from '../affix';
+import { Affix, AffixRef } from '../affix';
 import { ROW_LISTENERS } from './TR';
-import THead from './THead';
+import THead, { TheadProps } from './THead';
 import TFoot from './TFoot';
 import useTableHeader from './hooks/useTableHeader';
 import useColumnResize from './hooks/useColumnResize';
@@ -23,6 +32,7 @@ import { TableRowData } from './type';
 import useVirtualScroll from '../hooks/useVirtualScroll';
 import { getIEVersion } from '../_common/js/utils/helper';
 import log from '../_common/js/log';
+import useDefaultProps from '../hooks/useDefaultProps';
 
 export const BASE_TABLE_EVENTS = ['page-change', 'cell-click', 'scroll', 'scrollX', 'scrollY'];
 export const BASE_TABLE_ALL_EVENTS = ROW_LISTENERS.map((t) => `row-${t}`).concat(BASE_TABLE_EVENTS);
@@ -31,7 +41,8 @@ export interface TableListeners {
   [key: string]: Function;
 }
 
-const BaseTable = forwardRef<BaseTableRef, BaseTableProps>((props, ref) => {
+const BaseTable = forwardRef<BaseTableRef, BaseTableProps>((originalProps, ref) => {
+  const props = useDefaultProps<BaseTableProps<TableRowData>>(originalProps, baseTableDefaultProps);
   const {
     showHeader = true,
     tableLayout,
@@ -62,10 +73,10 @@ const BaseTable = forwardRef<BaseTableRef, BaseTableProps>((props, ref) => {
   );
 
   const { showElement } = useElementLazyRender(tableRef, lazyLoad);
-  const paginationAffixRef = useRef();
-  const horizontalScrollAffixRef = useRef();
-  const headerTopAffixRef = useRef();
-  const footerBottomAffixRef = useRef();
+  const paginationAffixRef = useRef<AffixRef>();
+  const horizontalScrollAffixRef = useRef<AffixRef>();
+  const headerTopAffixRef = useRef<AffixRef>();
+  const footerBottomAffixRef = useRef<AffixRef>();
 
   // 1. 表头吸顶；2. 表尾吸底；3. 底部滚动条吸底；4. 分页器吸底
   const {
@@ -284,7 +295,7 @@ const BaseTable = forwardRef<BaseTableRef, BaseTableProps>((props, ref) => {
       })}
     </colgroup>
   );
-  const headProps = {
+  const headProps: TheadProps = {
     isFixedHeader,
     rowAndColFixedPosition,
     isMultipleHeader,
@@ -727,10 +738,6 @@ const BaseTable = forwardRef<BaseTableRef, BaseTableProps>((props, ref) => {
 
 BaseTable.displayName = 'BaseTable';
 
-BaseTable.defaultProps = baseTableDefaultProps;
-
 export default BaseTable as <T extends TableRowData = TableRowData>(
-  props: BaseTableProps<T> & {
-    ref?: React.Ref<BaseTableRef>;
-  },
+  props: BaseTableProps<T> & RefAttributes<BaseTableRef>,
 ) => React.ReactElement;