+++ src/tag/Tag.tsx
@@ -1,12 +1,15 @@
-import React, { FocusEvent, forwardRef } from 'react';
+import React, { ForwardRefRenderFunction, FocusEvent, forwardRef, useMemo } from 'react';
 import classNames from 'classnames';
 import { CloseIcon as TdCloseIcon } from 'tdesign-icons-react';
+import tinycolor from 'tinycolor2';
+
 import noop from '../_util/noop';
 import useConfig from '../hooks/useConfig';
 import useGlobalIcon from '../hooks/useGlobalIcon';
 import { StyledProps } from '../common';
 import { TdTagProps } from './type';
 import { tagDefaultProps } from './defaultProps';
+import useDefaultProps from '../hooks/useDefaultProps';
 
 /**
  * Tag 组件支持的属性。
@@ -21,7 +24,8 @@ export interface TagProps extends TdTagProps, StyledProps {
   onBlur?: (e: FocusEvent<HTMLDivElement>) => void;
 }
 
-export function TagFunction(props: TagProps, ref: React.Ref<HTMLDivElement>) {
+export const TagFunction: ForwardRefRenderFunction<HTMLDivElement, TagProps> = (originalProps, ref) => {
+  const props = useDefaultProps<TagProps>(originalProps, tagDefaultProps);
   const {
     theme,
     size,
@@ -37,6 +41,7 @@ export function TagFunction(props: TagProps, ref: React.Ref<HTMLDivElement>) {
     style,
     disabled,
     children,
+    color,
     ...otherTagProps
   } = props;
 
@@ -83,6 +88,32 @@ export function TagFunction(props: TagProps, ref: React.Ref<HTMLDivElement>) {
   })();
   const titleAttribute = title ? { title } : undefined;
 
+  const getTagStyle = useMemo(() => {
+    if (!color) return style;
+    const luminance = tinycolor(color).getLuminance();
+
+    const calculatedStyle = style || {};
+
+    calculatedStyle.color = luminance > 0.5 ? 'black' : 'white';
+    if (variant === 'outline' || variant === 'light-outline') {
+      calculatedStyle.borderColor = color;
+    }
+
+    if (variant !== 'outline') {
+      const getLightestShade = () => {
+        const { r, g, b } = tinycolor(color).toRgb();
+        // alpha 0.1  is designed by @wen1kang
+        return `rgba(${r}, ${g}, ${b}, 0.1)`;
+      };
+
+      calculatedStyle.backgroundColor = variant === 'dark' ? color : getLightestShade();
+    }
+    if (variant !== 'dark') {
+      calculatedStyle.color = color;
+    }
+    return calculatedStyle;
+  }, [color, variant, style]);
+
   const tag = (
     <div
       ref={ref}
@@ -91,7 +122,9 @@ export function TagFunction(props: TagProps, ref: React.Ref<HTMLDivElement>) {
         if (disabled) return;
         onClick({ e });
       }}
-      style={maxWidth ? { maxWidth: typeof maxWidth === 'number' ? `${maxWidth}px` : maxWidth, ...style } : style}
+      style={
+        maxWidth ? { maxWidth: typeof maxWidth === 'number' ? `${maxWidth}px` : maxWidth, ...getTagStyle } : getTagStyle
+      }
       {...otherTagProps}
     >
       <>
@@ -105,11 +138,10 @@ export function TagFunction(props: TagProps, ref: React.Ref<HTMLDivElement>) {
   );
 
   return tag;
-}
+};
 
 export const Tag = forwardRef(TagFunction);
 
 Tag.displayName = 'Tag';
-Tag.defaultProps = tagDefaultProps;
 
 export default Tag;