+++ src/tooltip/Tooltip.tsx
@@ -1,13 +1,15 @@
 import React, { forwardRef, useState, useEffect, useRef, useImperativeHandle } from 'react';
 import classNames from 'classnames';
-import Popup, { PopupVisibleChangeContext } from '../popup';
+import Popup, { PopupRef, PopupVisibleChangeContext } from '../popup';
 import useConfig from '../hooks/useConfig';
 import { TdTooltipProps } from './type';
 import { tooltipDefaultProps } from './defaultProps';
+import useDefaultProps from '../hooks/useDefaultProps';
 
 export type TooltipProps = TdTooltipProps;
 
-const Tooltip = forwardRef((props: TdTooltipProps, ref) => {
+const Tooltip = forwardRef<Partial<PopupRef>, TdTooltipProps>((originalProps, ref) => {
+  const props = useDefaultProps<TdTooltipProps>(originalProps, tooltipDefaultProps);
   const {
     theme,
     showArrow,
@@ -22,7 +24,7 @@ const Tooltip = forwardRef((props: TdTooltipProps, ref) => {
 
   const { classPrefix } = useConfig();
   const [timeUp, setTimeUp] = useState(false);
-  const popupRef = useRef(null);
+  const popupRef = useRef<PopupRef>(null);
   const timerRef = useRef<number | null>(null);
   const toolTipClass = classNames(
     `${classPrefix}-tooltip`,
@@ -51,7 +53,7 @@ const Tooltip = forwardRef((props: TdTooltipProps, ref) => {
   }, [duration, timeUp]);
 
   useImperativeHandle(ref, () => ({
-    ...((popupRef.current || {}) as any),
+    ...(popupRef.current || {}),
   }));
 
   return (
@@ -70,6 +72,5 @@ const Tooltip = forwardRef((props: TdTooltipProps, ref) => {
 });
 
 Tooltip.displayName = 'Tooltip';
-Tooltip.defaultProps = tooltipDefaultProps;
 
 export default Tooltip;