+++ src/tree/Tree.tsx
@@ -28,11 +28,21 @@ import useTreeVirtualScroll from './hooks/useTreeVirtualScroll';
 
 import type { TreeNodeState, TreeNodeValue, TypeTreeNodeData, TypeTreeNodeModel } from '../_common/js/tree-v1/types';
 import type { TreeInstanceFunctions, TdTreeProps } from './type';
+import useDefaultProps from '../hooks/useDefaultProps';
 
 export type TreeProps = TdTreeProps & StyledProps;
 
-const Tree = forwardRef((props: TreeProps, ref: React.Ref<TreeInstanceFunctions>) => {
+const Tree = forwardRef<TreeInstanceFunctions<TreeOptionData>, TreeProps>((originalProps, ref) => {
   const { treeClassNames, transitionNames, transitionClassNames, transitionDuration, locale } = useTreeConfig();
+  const props = useDefaultProps<TreeProps>(originalProps, {
+    data: [],
+    expandLevel: 0,
+    icon: true,
+    line: false,
+    transition: true,
+    lazy: true,
+    valueMode: 'onlyLeaf',
+  });
 
   // 可见节点集合
   const [visibleNodes, setVisibleNodes] = useState([]);
@@ -364,15 +374,5 @@ const Tree = forwardRef((props: TreeProps, ref: React.Ref<TreeInstanceFunctions>
 });
 
 Tree.displayName = 'Tree';
-
-Tree.defaultProps = {
-  data: [],
-  expandLevel: 0,
-  icon: true,
-  line: false,
-  transition: true,
-  lazy: true,
-  valueMode: 'onlyLeaf',
-};
 
 export default Tree;